requirements.txt
aws-cdk-lib==2.86.0
constructs>=10.0.0,<11.0.0
boto3
boto3-stubs[essential]
boto3-stubs[stepfunctions]
boto3-stubs[sagemaker-a2i-runtime]
boto3-stubs[cloudwatch]
boto3-stubs[logs]
jsii
DotMap
Pillow
XlsxWriter
pandas
pytz
uuid
pyrhubarb
docker